feat: test new retriever with string parser (not working)
Added new string parser to the pipe to pass the question as  a string instead of an object.
Tried to run the model but an issue is present in regard of the vector embeddings size
it is 4096 instead of the maximum allowed by MongoDB Atlas search index's. This seems
to be related to hte DB used (Mongo) and/or by the Embedder used (OllamaEmbeddings)
I will try to come up with a solution to this by either changind vector storage db
or by changing embedder.